% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoleverage.R
\name{algoleverage}
\alias{algoleverage}
\title{algoleverage Function}
\usage{
algoleverage(method, X, Y, n, r, Pi = NULL, replication = 500, beta_ols)
}
\arguments{
\item{method}{The method for algorithmic leveraging, choose from "unif","leverage".}

\item{X}{dataset X}

\item{Y}{dataset Y}

\item{n}{sample size of data set}

\item{r}{size of a randomly selected subset}

\item{Pi}{sampling probability}

\item{replication}{replication, how many draws of subsample}

\item{beta_ols}{the linear regression coefficient from ols}
}
\value{
return the absolute error between coefficient from leveraging algorithm and coefficient from ols for each draw of subsample
}
\description{
This function implements algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows.
}
\examples{
set.seed(65206520)
n=500
X=rt(n, 6)
eps=rnorm(n,0,1)
Y=-1*X+eps
beta_ols=lm(Y ~ 0 + X )$coefficients
algoleverage("unif",X,Y,n,r=10,replication=500,beta_ols=beta_ols)
}
