% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnetcoord.R
\name{elasticnet}
\alias{elasticnet}
\title{elasticnet Function}
\usage{
elasticnet(X, Y, alpha, n, p, nlambda = 100, tol = 1e-05)
}
\arguments{
\item{X}{dataset X}

\item{Y}{dataset Y}

\item{alpha}{the parameter determining the ratio of L1 penalty to add}

\item{n}{sample size of data set}

\item{p}{dimension}

\item{nlambda}{the length of the sequence of the lambda}

\item{tol}{tolerance; when the difference less than tolerance, stop the algorithm}
}
\value{
return a list of estimated coefficients for each lambda
}
\description{
This function fits elastic net to data using coordinate descent algorithm.
}
\examples{
p=20
n=20
mu <- rep(0,20)
Sigma <- diag(rep(1,20))
Sigma[1,2]=0.8
Sigma[2,1]=0.8
Sigma[5,6]=0.8
Sigma[6,5]=0.8
X=matrix(NA,n,p)
set.seed(65206520)
for (i in 1:n){X[i,]=MASS::mvrnorm(n=1, mu=mu, Sigma=Sigma)}
beta=c(2,0,-2,0,1,0,-1,0,rep(0,12))
set.seed(65206520)
eps=rnorm(n,0,1)
Y=X\%*\%beta+eps
elasticnet(X,Y,alpha=0,n,p,nlambda=100,tol=1e-5)
}
\seealso{
\link[glmnet]{glmnet}
}
