% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveols.R
\name{solveols}
\alias{solveols}
\title{solveols Function}
\usage{
solveols(method, cores = NULL, A, v, iter = 10000)
}
\arguments{
\item{method}{The method to solve a linear system, choose from "GaussSeidel","Jacobi"}

\item{cores}{Defaults to NULL, how many cores to use for parallel implementation.}

\item{A}{is n×n tridiagonal with alpha on the diagonals and -1 on theimmediate off diagonals.}

\item{v}{a vector repeated with 1 and 0.}

\item{iter}{the maximum iteration}
}
\value{
return a list of error and runtime for each iteration
}
\description{
This function solves a linear system using Gauss-Seidel or Jacobi method, allows user to specify how many cores to use for parallel implementation.
}
\examples{
alpha3=3
n <- 100
A3 <- diag(alpha3, n)
A3[abs(row(A3) - col(A3)) == 1] <- -1
v=rep(c(1,0),n/2)
solveols("GaussSeidel",A=A3,v=v,iter=10)


}
